Update sources - Add ext2 initial write support - Improved vfs dcache - Fixed pf on userspace with ap cpus - Fixed issues

+++ arch/x86-family/cpu/cpu.c
@@ -43,7 +43,7 @@
 
 
 // see startup.S
-extern int startup_tss;
+extern uint8_t startup_tss;
 
 
 
@@ -67,7 +67,7 @@ void arch_cpu_init(int index) {
     for(i = 0; i < SMP_CPU_MAX_FEATURES; i++)
         core->cpu.cores[index].features[i] = 0ULL;
         
-    core->cpu.cores[index].tss = (void*) (&startup_tss + (index * sizeof(tss_t)));
+    core->cpu.cores[index].tss = (void*) (((uintptr_t) &startup_tss) + (index * sizeof(tss_t)));
 
 
     spinlock_init_with_flags(&core->cpu.cores[index].global_lock, SPINLOCK_FLAGS_CPU_OWNER);
@@ -404,41 +404,41 @@ void arch_cpu_init(int index) {
 
 
 
-    //! Enable TSC Timer
+    //? Enable TSC Timer
     if(cpu_has(index, X86_FEATURE_TSC))
         x86_set_cr4(x86_get_cr4() | X86_CR4_TSD_MASK);
 
 
-    //! Enable NX bit
+    //? Enable NX bit
     if(cpu_has(index, X86_FEATURE_NX))
         x86_wrmsr(X86_MSR_EFER, x86_rdmsr(X86_MSR_EFER) | X86_MSR_EFER_NXE);
 
 
-    //! Enable Syscall bit
+    //? Enable Syscall bit
     if(cpu_has(index, X86_FEATURE_SYSCALL))
         x86_wrmsr(X86_MSR_EFER, x86_rdmsr(X86_MSR_EFER) | X86_MSR_EFER_SCE);
 
 
-    //! Enable FSGSBASE instructions
+    //? Enable FSGSBASE instructions
     if(cpu_has(index, X86_FEATURE_FSGSBASE))
         x86_set_cr4(x86_get_cr4() | X86_CR4_FSGSBASE_MASK);
 
 
 #if defined(CONFIG_X86_ENABLE_SMEP)
-    //! Enable SMEP
+    //? Enable SMEP
     if(cpu_has(index, X86_FEATURE_SMEP))
         x86_set_cr4(x86_get_cr4() | X86_CR4_SMEP_MASK);
 #endif
 
 #if defined(CONFIG_X86_ENABLE_SMAP)
-    // //! Enable SMAP
+    //? Enable SMAP
     if(cpu_has(index, X86_FEATURE_SMAP))
         x86_set_cr4(x86_get_cr4() | X86_CR4_SMAP_MASK);
 #endif
 
 
 #if defined(CONFIG_HAVE_SMP)
-    //! Write Processor ID
+    //? Write Processor ID
     if(cpu_has(index, X86_FEATURE_RDTSCP))
         x86_wrmsr(X86_MSR_TSC_AUX, index);
 #endif
